Moved functions into utils
